saalN-hd-direct: add icedist mode
This also modernizes the ffmpeg input handling (using typed stream
specifiers) and also pushes through the 2nd translation channel in any
case. However, this is only used when running in icedist mode, as RTMP
cannot contain three audio channels.

We also switch from using mpegts for the mux to matroska, as the icedist
sink (using matroska as a container itself) requires the codec headers
to be present early, which doesn't seem to be working with mpegts.

Alas, this means that whenever a sink fails, the whole sdi-script
profile needs to be restarted, as mkv is not self-synchronizing.