register2ansible: add nginx variable hacks
Always define all master variables because the config templates don't
handle undefinedness (for example on an Icecast-only relay which only
needs nginx for SSL-proxying). Last year, this probably wasn't a problem
because every relay was doing everything.

HTTP caching stuff is free as long as we don't push clients on there,
which we'll handle in the load balancer.