saalN-hd-direct: drop RTMP multiaudio support
Not sure if this was actually causing problems, but we dropped this
during debugging of the volume audiofilter.

This was only present for fallback from the icedist stuff, which seems
to be running stably enough, so we drop it to align the config on the
boxes to the one in Ansible.