ansible-playbook-keepass: use python2
When not using a virtualenv(2) to run this script, as of 68a7ca3, we use
the "python" binary (instead of a hardcoded path to a "python2" binary,
as the shebang of "ansible-playbook" specifies). On distributions such
as Archlinux, which symlink "python" to Python 3, this will fail.

As virtualenvs also provide a "python2" symlink, we can simply use that
and hopefully work properly on all distros.