fanout/h264: use variant HLS streams
Use variant HLS streams for translation tracks to prevent duplication of
video streams in the CDN.

In addition, we now provide resolution-specific master files both for
backwards compatibility and user choice. Those still do adaptive bitrate
switching, but only up to the limit specified in the filename. For
example, this allows users on a metered connection to choose a
lower-quality stream even if their connection can handle a higher
bitrate.