check_system: make hypervisor check more robust
Currently, we check whether we are running under some kind of
virtualization by grepping in the dmesg output. This ceases to work if
the dmesg ring buffer starts losing the early boot output and can cause
spurious problems to appear on long-running machines.

As we're running systemd everywhere nowadays, we can just use
systemd-detect-virt for this purpose, which is much more robust.